# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AFRICA CDC KnowledgeHub API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/knowledge_hub",
            "description": "Current deployment (set APP_URL in .env)"
        }
    ],
    "paths": {
...
        "/api/forums": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "List Forums",
                "description": "Returns a list of all Forums",
                "operationId": "ListForums",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Search term for specific records",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "community_id",
                        "in": "query",
                        "description": "Filter by Community Id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Create Forum",
                "description": "Allows users to submit Forum Discussion",
                "operationId": "CreateForum",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Photo file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request, when some required data is missing"
                    },
                    "404": {
                        "description": "Not found when you send the request to an invalid endpoint"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/forums/comment": {
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Create Forum Comment",
                "description": "Allows users to submit Forum Discussion Comment",
                "operationId": "CreateForumComment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request, when some required data is missing"
                    },
                    "404": {
                        "description": "Not found when you send the request to an invalid endpoint"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/forums/{id}": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Show Forum",
                "description": "Retrieve a single forum record based on the ID.",
                "operationId": "ShowForum",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Forum ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}